+++ allshoppings-services/src/main/java/mobi/allshoppings/dao/APDMABlackListDAO.java
@@ -13,6 +13,7 @@ import mobi.allshoppings.tools.Range;
 public interface APDMABlackListDAO extends GenericDAO<APDMABlackList> {
 
 	Key createKey() throws ASException;
-	List<APDMABlackList> getUsingEntityIdAndRange(String entityId, Integer entityKind, Range range, String order, Map<String, String> attributes, boolean detachable) throws ASException;
+	List<APDMABlackList> getUsingEntityIdAndRange(String entityId, byte entityKind, Range range,
+			String order, Map<String, String> attributes, boolean detachable) throws ASException;
 
 }
