+++ allshoppings-services/src/main/java/mobi/allshoppings/dao/APDMAEmployeeDAO.java
@@ -13,7 +13,8 @@ import mobi.allshoppings.tools.Range;
 public interface APDMAEmployeeDAO extends GenericDAO<APDMAEmployee> {
 
 	Key createKey() throws ASException;
-	List<APDMAEmployee> getUsingEntityIdAndRange(String entityId, Integer entityKind, Range range, String order, Map<String, String> attributes, boolean detachable) throws ASException;
-	List<APDMAEmployee> getUsingEntityIdandMac(String entityId, Integer entityKind, String mac) throws ASException;
+	List<APDMAEmployee> getUsingEntityIdAndRange(String entityId, byte entityKind, Range range, String order,
+			Map<String, String> attributes, boolean detachable) throws ASException;
+	List<APDMAEmployee> getUsingEntityIdandMac(String entityId, byte entityKind, String mac) throws ASException;
 
 }
